requirements.txt
# Core Framework
langchain>=0.1.0
langchain-openai>=0.1.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Additional LangChain components
langchain-community>=0.0.10
langchain-core>=0.1.0
